Refactor test app to split App into feature modules and shared components
