+++ apps/factory-test-app/components/index.ts
@@ -1,0 +1,1 @@
+export { EventLog } from './EventLog';
