+++ apps/factory-test-app/features/app-shell/index.ts
@@ -1,0 +1,2 @@
+export { Lifecycle } from './Lifecycle';
+export { Stores } from './Stores';
